Refactor io module
