+++ python/lsst/scarlet/lite/io/blend_base.py
@@ -1,0 +1,61 @@
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from typing import ClassVar, Any
+
+from numpy.typing import DTypeLike
+
+from .errors import PersistenceError
+
+__all__ = ["ScarletBlendBaseData"]
+
+
+@dataclass(kw_only=True)
+class ScarletBlendBaseData(ABC):
+    """Base data for a scarlet Blend data."""
+
+    blend_registry: ClassVar[dict[str, type[ScarletBlendBaseData]]] = {}
+    blend_type: str
+    metadata: dict[str, Any] | None = None
+    version: str
+
+    @classmethod
+    def register(cls) -> None:
+        """Register a new Blend type"""
+        ScarletBlendBaseData.blend_registry[cls.blend_type] = cls
+
+    @abstractmethod
+    def as_dict(self) -> dict:
+        """Return the object encoded into a dict for JSON serialization
+
+        Returns
+        -------
+        result :
+            The object encoded as a JSON compatible dict
+        """
+
+    @staticmethod
+    def from_dict(data: dict, dtype: DTypeLike | None = None) -> ScarletBlendBaseData:
+        """Reconstruct `ScarletBlendBaseData` from JSON compatible dict.
+
+        Parameters
+        ----------
+        data :
+            Dictionary representation of the object
+        dtype :
+            Datatype of the resulting model.
+
+        Returns
+        -------
+        result :
+            The reconstructed object
+        """
+        # Default to "blend" for backwards compatibility
+        blend_type = data.get("blend_type", "blend")
+
+        if blend_type not in ScarletBlendBaseData.blend_registry:
+            raise PersistenceError(f"Unknown blend type: {blend_type}")
+
+        cls = ScarletBlendBaseData.blend_registry[blend_type]
+        return cls.from_dict(data, dtype=dtype)
