+++ python/lsst/scarlet/lite/io/errors.py
@@ -1,0 +1,6 @@
+__all__ = ["PersistenceError"]
+
+
+class PersistenceError(Exception):
+    """Custom error for persistence issues."""
+    pass
