+++ tests/test_component.py
@@ -19,6 +19,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+from __future__ import annotations
+
 from typing import Callable
 
 import numpy as np
@@ -46,6 +48,9 @@ class DummyComponent(Component):
 
     def parameterize(self, parameterization: Callable) -> None:
         parameterization(self)
+
+    def to_component_data(self) -> DummyComponent:
+        pass
 
 
 class TestFactorizedComponent(ScarletTestCase):
